pyproject.toml
# pyproject.toml

[build-system]
requires      = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "PyFibreBundle"
version = "1.3"
description = "Image processing of images acquired through fibre imaging bundle, including core removal, mosaicing and super-resolution.."
readme = "README.md"
authors = [{ name = "Michael Hughes", email = "m.r.hughes@kent.ac.uk" }]
license = { file = "LICENSE" }
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
]
keywords = ["fibre", "fiber", "bundle", "endomicroscopy", "endoscopic microscopy", "imaging"]
dependencies = [
    "matplotlib>=3.3.4",
    "numba>=0.55.1",
    "numpy >= 1.22",
    "opencv_python>=4.5",
    "Pillow>=9",
    "scipy>=1.7"   
]
requires-python = ">=3.8"

[project.urls]
Homepage = "https://github.com/MikeHughesKent/PyFibreBundle/"

